modified

+++ todo/views.py
@@ -1,8 +1,37 @@
 from django.shortcuts import render,redirect
+from django.urls import reverse_lazy,reverse
 from django.http import HttpResponse
 from .models import TodoList
 from .forms import TodoForm
+from django.views.generic import ListView
+from django.views.generic.detail import DetailView
+from django.views.generic.edit import UpdateView,DeleteView
+
 # Create your views here.
+
+class Tasklistview(ListView):
+	model=TodoList
+	template_name='add.html'
+	context_object_name='task'
+
+class TaskDetailview(DetailView):
+	model=TodoList
+	template_name='detail.html'
+	context_object_name='task'
+
+class Taskupdateview(UpdateView):
+	model=TodoList
+	template_name='edit.html'
+	context_object_name='task'
+	fields=('name','priority','date')
+	def get_success_url(self):
+		return reverse_lazy('cbvdetail',kwargs={'pk':self.object.id})
+
+class TaskDeleteView(DeleteView):
+	model=TodoList
+	template_name='delete.html'
+	success_url=reverse_lazy('cbvhome')
+
 def add_task(request):
 	if request.method=='POST':
 		name=request.POST.get('name')
@@ -17,9 +46,12 @@ def add_task(request):
 		return render(request,'add.html',{'task':task})
 	
 def delete(request,task_id):
-	todo=TodoList.objects.get(id=task_id)
-	todo.delete()
-	return redirect('/')
+	if request.method=='POST':
+		todo=TodoList.objects.get(id=task_id)
+		todo.delete()
+		return redirect('/')
+	else:
+		return render(request,'delete.html')
 
 def update(request,task_id):
 	task=TodoList.objects.get(id=task_id)
